created a specific file to host FlickStick class
